+++ crates/biome_js_analyze/src/analyzers/nursery/use_shorthand_assign.rs
@@ -1,0 +1,210 @@
+use biome_analyze::{
+    context::RuleContext, declare_rule, ActionCategory, Ast, Rule, RuleDiagnostic,
+};
+use biome_console::markup;
+use biome_diagnostics::Applicability;
+use biome_js_factory::make;
+use biome_js_syntax::{
+    AnyJsAssignment, AnyJsExpression, JsAssignmentExpression, JsAssignmentOperator,
+    JsBinaryExpression, JsBinaryOperator, JsSyntaxKind, T,
+};
+use biome_rowan::{AstNode, BatchMutationExt};
+
+use crate::JsRuleAction;
+
+declare_rule! {
+    /// Require assignment operator shorthand where possible.
+    ///
+    /// JavaScript provides shorthand operators combining a variable assignment and simple mathematical operation.
+    ///
+    /// Source: https://eslint.org/docs/latest/rules/operator-assignment/
+    ///
+    /// ## Examples
+    ///
+    /// ### Invalid
+    ///
+    /// ```js,expect_diagnostic
+    /// a = a + 1;
+    /// ```
+    ///
+    /// ```js,expect_diagnostic
+    /// a = a - 1;
+    /// ```
+    ///
+    ///  ```js,expect_diagnostic
+    /// a = a * 1;
+    /// ```
+    ///
+    /// ## Valid
+    ///
+    /// ```js
+    /// a += 1;
+    /// ```
+    ///
+    /// ```js
+    /// a -= 1;
+    /// ```
+    ///
+    ///  ```js
+    /// a *= 1;
+    /// ```
+    ///
+    pub(crate) UseShorthandAssign {
+        version: "next",
+        name: "useShorthandAssign",
+        recommended: false,
+    }
+}
+
+pub struct RuleState {
+    shorthand_operator: JsSyntaxKind,
+    replacement_expression: AnyJsExpression,
+}
+
+#[derive(Clone)]
+enum VariablePosition {
+    Left,
+    Right,
+}
+
+impl Rule for UseShorthandAssign {
+    type Query = Ast<JsAssignmentExpression>;
+    type State = RuleState;
+    type Signals = Option<Self::State>;
+    type Options = ();
+
+    fn run(ctx: &RuleContext<Self>) -> Self::Signals {
+        let node = ctx.query();
+
+        if !matches!(node.operator(), Ok(JsAssignmentOperator::Assign)) {
+            return None;
+        }
+
+        let left = node.left().ok()?;
+        let right = node.right().ok()?;
+
+        let left_var_name = match left.as_any_js_assignment()? {
+            AnyJsAssignment::JsComputedMemberAssignment(assignment) => assignment.text(),
+            AnyJsAssignment::JsIdentifierAssignment(assignment) => assignment.text(),
+            AnyJsAssignment::JsStaticMemberAssignment(assignment) => assignment.text(),
+            _ => return None,
+        };
+
+        let binary_expression = match right {
+            AnyJsExpression::JsBinaryExpression(binary_expression) => binary_expression,
+            AnyJsExpression::JsParenthesizedExpression(param) => {
+                JsBinaryExpression::cast_ref(param.expression().ok()?.syntax())?
+            }
+            _ => return None,
+        };
+
+        let operator = binary_expression.operator().ok()?;
+        let shorthand_operator = get_shorthand(operator)?;
+        let is_commutative = is_commutative(operator);
+        let variable_position_in_expression =
+            parse_variable_reference_in_expression(&left_var_name, &binary_expression)?;
+
+        let replacement_expression = match variable_position_in_expression {
+            VariablePosition::Left => binary_expression.left().ok()?,
+            VariablePosition::Right => binary_expression.right().ok()?,
+        };
+
+        if !is_commutative && matches!(variable_position_in_expression, VariablePosition::Right) {
+            return None;
+        }
+
+        Some(RuleState {
+            shorthand_operator,
+            replacement_expression,
+        })
+    }
+
+    fn diagnostic(ctx: &RuleContext<Self>, state: &Self::State) -> Option<RuleDiagnostic> {
+        let node = ctx.query();
+        let shorthand_operator = state.shorthand_operator.to_string()?;
+
+        Some(RuleDiagnostic::new(
+            rule_category!(),
+            node.range(),
+            markup! {
+                "Assignment "<Emphasis>"(=)"</Emphasis>" can be replaced with operator assignment "<Emphasis>""{shorthand_operator}""</Emphasis>"."
+            },
+        ))
+    }
+
+    fn action(ctx: &RuleContext<Self>, state: &Self::State) -> Option<JsRuleAction> {
+        let node = ctx.query();
+
+        let mut mutation = ctx.root().begin();
+        let shorthand_operator = state.shorthand_operator;
+
+        let token_leading_trivia = node.operator_token().ok()?.leading_trivia().pieces();
+        let token_trailing_trivia = node.operator_token().ok()?.trailing_trivia().pieces();
+        let token = make::token(shorthand_operator)
+            .with_leading_trivia_pieces(token_leading_trivia)
+            .with_trailing_trivia_pieces(token_trailing_trivia);
+
+        let shorthand_node = node
+            .clone()
+            .with_operator_token_token(token)
+            .with_right(state.replacement_expression.clone());
+
+        mutation.replace_node(node.clone(), shorthand_node);
+
+        Some(JsRuleAction {
+            category: ActionCategory::QuickFix,
+            applicability: Applicability::MaybeIncorrect,
+            message: markup! { "Use "<Emphasis>""{shorthand_operator.to_string()?}""</Emphasis>" instead." }
+                .to_owned(),
+            mutation,
+        })
+    }
+}
+
+fn parse_variable_reference_in_expression(
+    variable_name: &str,
+    binary_expression: &JsBinaryExpression,
+) -> Option<VariablePosition> {
+    let present_on_left = variable_name == binary_expression.left().ok()?.omit_parentheses().text();
+
+    if present_on_left {
+        return Some(VariablePosition::Left);
+    }
+
+    let present_on_right =
+        variable_name == binary_expression.right().ok()?.omit_parentheses().text();
+
+    if present_on_right {
+        Some(VariablePosition::Right)
+    } else {
+        None
+    }
+}
+
+fn is_commutative(operator: JsBinaryOperator) -> bool {
+    matches!(
+        operator,
+        JsBinaryOperator::Times
+            | JsBinaryOperator::BitwiseAnd
+            | JsBinaryOperator::BitwiseOr
+            | JsBinaryOperator::BitwiseXor
+    )
+}
+
+fn get_shorthand(operator: JsBinaryOperator) -> Option<JsSyntaxKind> {
+    match operator {
+        JsBinaryOperator::Plus => Some(T![+=]),
+        JsBinaryOperator::Minus => Some(T![-=]),
+        JsBinaryOperator::Times => Some(T![*=]),
+        JsBinaryOperator::Divide => Some(T![/=]),
+        JsBinaryOperator::Remainder => Some(T![%=]),
+        JsBinaryOperator::Exponent => Some(T![**=]),
+        JsBinaryOperator::BitwiseAnd => Some(T![&=]),
+        JsBinaryOperator::BitwiseOr => Some(T![|=]),
+        JsBinaryOperator::BitwiseXor => Some(T![^=]),
+        JsBinaryOperator::LeftShift => Some(T![<<=]),
+        JsBinaryOperator::RightShift => Some(T![>>=]),
+        JsBinaryOperator::UnsignedRightShift => Some(T![>>>=]),
+        _ => None,
+    }
+}