+++ crates/biome_js_analyze/tests/specs/nursery/useShorthandAssign/valid.js
@@ -1,0 +1,17 @@
+/* should not generate diagnostics */
+
+a += 1;
+a -= 1;
+a *= 1;
+a /= 1;
+a %= 1;
+a <<= 1;
+a >>= 1;
+a >>>= 1;
+a &= 1;
+a ^= 1;
+a |= 1;
+a **= 1;
+this.a += 1;
+x[0] -= x[0]
+window.a += window.a;