+++ crates/biome_service/src/configuration/parse/json/rules.rs
@@ -2042,6 +2042,7 @@ impl VisitNode<JsonLanguage> for Nursery {
                 "useHookAtTopLevel",
                 "useImportRestrictions",
                 "useIsArray",
+                "useShorthandAssign",
             ],
             diagnostics,
         )
@@ -2490,6 +2491,29 @@ impl VisitNode<JsonLanguage> for Nursery {
                         diagnostics,
                     )?;
                     self.use_is_array = Some(rule_configuration);
+                }
+                _ => {
+                    diagnostics.push(DeserializationDiagnostic::new_incorrect_type(
+                        "object or string",
+                        value.range(),
+                    ));
+                }
+            },
+            "useShorthandAssign" => match value {
+                AnyJsonValue::JsonStringValue(_) => {
+                    let mut configuration = RuleConfiguration::default();
+                    self.map_to_known_string(&value, name_text, &mut configuration, diagnostics)?;
+                    self.use_shorthand_assign = Some(configuration);
+                }
+                AnyJsonValue::JsonObjectValue(_) => {
+                    let mut rule_configuration = RuleConfiguration::default();
+                    rule_configuration.map_rule_configuration(
+                        &value,
+                        name_text,
+                        "useShorthandAssign",
+                        diagnostics,
+                    )?;
+                    self.use_shorthand_assign = Some(rule_configuration);
                 }
                 _ => {
                     diagnostics.push(DeserializationDiagnostic::new_incorrect_type(