+++ packages/@biomejs/backend-jsonrpc/src/workspace.ts
@@ -745,6 +745,10 @@ export interface Nursery {
 	 * Use Array.isArray() instead of instanceof Array.
 	 */
 	useIsArray?: RuleConfiguration;
+	/**
+	 * Require assignment operator shorthand where possible.
+	 */
+	useShorthandAssign?: RuleConfiguration;
 }
 /**
  * A list of rules that belong to this group
@@ -1315,6 +1319,7 @@ export type Category =
 	| "lint/nursery/useHookAtTopLevel"
 	| "lint/nursery/useImportRestrictions"
 	| "lint/nursery/useIsArray"
+	| "lint/nursery/useShorthandAssign"
 	| "lint/performance/noDelete"
 	| "lint/security/noDangerouslySetInnerHtml"
 	| "lint/security/noDangerouslySetInnerHtmlWithChildren"